% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_score.R
\name{calc_score}
\alias{calc_score}
\title{Calculate scores based on LASSO coefficient}
\usage{
calc_score(GEO, lassocoef = lassocoef, fd = NULL, is.count = FALSE,
  log = FALSE, destdir = tempdir())
}
\arguments{
\item{GEO}{character, GEO accession number}

\item{lassocoef}{data.frame, table of coefficient, result from lass_cox}

\item{fd}{data.frame (optional), feature data including id and gene symbol mapping}

\item{is.count}{logical, default is FALSE. is the GEO data count matrix (RNA-seq) or log transformed expression (microarray or RNA-seq)}

\item{log}{logical, default is FALSE, whether or not need log transform. Depends on the normalization of expression matrix}

\item{destdir}{where to store downloaded GEO data, default is tempdir()}
}
\value{
List of data, include scores for each sample, feature data, pheno data, expression matrix for all and subset genes
}
\description{
Calculate scores based on LASSO coefficient using data from GEO
}
\examples{
lassocoef <- lasso_cox(exp, feature, surv)
score <- calc_score("GSE7440", lassocoef, destdir = ".", log = TRUE)

}
