% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_scores_simply.R
\name{calc_scores_simply}
\alias{calc_scores_simply}
\title{Calculate scores based on LASSO coefficient_ simplified}
\usage{
calc_scores_simply(exp, fd, lassocoef)
}
\arguments{
\item{exp}{matrix, expression matrix}

\item{fd}{data.frame, feature data including id and gene symbol mapping}

\item{lassocoef}{data.frame, table of coefficient, result from lass_cox}
}
\value{
matrix, rownames are sample ids, first coloumn is score
}
\description{
Calculate scores based on LASSO coefficient using expression matrix
}
\examples{
lassocoef <- lasso_cox(exp, feature, surv)
score <- calc_score("GSE7440", lassocoef, destdir = ".", log = TRUE)
score2 <- calc_scores_simply(score$exp, score$feature, lassocoef)

}
