% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exp_data.R
\docType{data}
\name{exp}
\alias{exp}
\title{RNA-seq gene expression from TARGET phase II}
\format{A data frame with 137 rows and 24918 variables:}
\usage{
exp
}
\description{
RNA-seq gene expression from TARGET phase II
}
\keyword{datasets}
