% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lasso_cox.R
\name{lasso_cox}
\alias{lasso_cox}
\title{generate LASSO-Cox model using input data}
\usage{
lasso_cox(exp, feature, surv, lambda = FALSE)
}
\arguments{
\item{exp}{data.frame, expression matrix, log transformed, first column is patient id (PID), other column is gene}

\item{feature}{vector, genes names, consitent with expression data frame colnames}

\item{surv}{data.frame, include columnd os.time and os.status,first column must be identical as first column in exp}

\item{lambda}{logical, whether output the lambda min value}
}
\value{
List: coefficient and lambda (if required)
}
\description{
Use input expression, clinical and gene list to generate LASSO-Cox model
}
\examples{
lassocoef <- lasso_cox(exp, feature, surv, lambda = FALSE)

}
