% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surv_data.R
\docType{data}
\name{surv}
\alias{surv}
\title{clinical information from TARGET phase II}
\format{A data frame with 137 rows and 3 variables:
\describe{
  \item{clinical..TARGET.USI.}{patient ID}
  \item{os.status}{overall survival status, 1 is dead, 0 is alive}
  \item{os.time}{overall survival time}
}}
\usage{
surv
}
\description{
clinical information from TARGET phase II
}
\keyword{datasets}
