% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topN.R
\name{topN}
\alias{topN}
\title{generate LASSO-Cox model based on occurence in permutations}
\usage{
topN(x, f, y, perm, n = 3)
}
\arguments{
\item{x}{data.frame, expression matrix, log transformed, first column is patient id (PID), other column is gene}

\item{f}{vector, genes names, consitent with expression data frame colnames}

\item{y}{data.frame, include columnd os.time and os.status,first column must be identical as first column in exp}

\item{perm}{numeric, number of permutations}

\item{n}{numeric, top n models will be output}
}
\value{
List: coefficient and lambda (if required)
}
\description{
Use input expression, clinical and gene list to generate LASSO-Cox model, permutate and obtain most often occured model
}
\examples{
\dontrun{
top <- topN(exp, feature, surv, perm=100, n=3)
}

}
