% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lasso_cox.R
\name{lasso_cox}
\alias{lasso_cox}
\title{generate LASSO-Cox model using input data}
\usage{
lasso_cox(exp, feature, surv, lambda = FALSE)
}
\arguments{
\item{exp}{matrix, expression matrix, log transformed, row is patient, column is gene}

\item{feature}{vector, genes, consitent with expression matrix colnames}

\item{survdata}{data.frame, include columnd os.time and os.status, rownames same as rowname in expression matrix}
}
\value{
List: coefficient and lambda (if required)
}
\description{
Use input expression, clinical and gene list to generate LASSO-Cox model
}
\examples{
lassocoef <- lasso_cox(x, f, y, lambda = T)

}
