% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topN.R
\name{topN}
\alias{topN}
\title{generate LASSO-Cox model based on occurence in permutations}
\usage{
topN(x, f, y, perm, n = 3)
}
\arguments{
\item{x}{matrix, expression matrix, log transformed, row is patient, column is gene}

\item{f}{vector, genes, consitent with expression matrix colnames}

\item{y}{data.frame, include columnd os.time and os.status, rownames same as rowname in expression matrix}

\item{perm}{numeric, number of permutations}
}
\value{
List: coefficient and lambda (if required)
}
\description{
Use input expression, clinical and gene list to generate LASSO-Cox model, permutate and obtain most often occured model
}
\examples{
top <- topN(x, f, y, perm=100, n=3)

}
